- Moved Model to Entity transitions to Mappers
