+++ InnopolisUniversity/AndroidDevelopment/MyApp/app/src/main/java/me/igorfedorov/myapp/feature/weather_screen/data/api/model/CoordinatesModel.kt
@@ -1,7 +1,6 @@
 package me.igorfedorov.myapp.feature.weather_screen.data.api.model
 
 import com.google.gson.annotations.SerializedName
-import me.igorfedorov.myapp.feature.weather_screen.domain.model.Coordinates
 
 data class CoordinatesModel(
     @SerializedName("lon")
@@ -9,10 +8,3 @@ data class CoordinatesModel(
     @SerializedName("lat")
     val lat: Double
 )
-
-fun CoordinatesModel.toCoordinates(): Coordinates {
-    return Coordinates(
-        lon = lon,
-        lat = lat
-    )
-}
