+++ InnopolisUniversity/AndroidDevelopment/MyApp/app/src/main/java/me/igorfedorov/myapp/feature/weather_screen/data/api/model/WeatherMainModel.kt
@@ -1,7 +1,6 @@
 package me.igorfedorov.myapp.feature.weather_screen.data.api.model
 
 import com.google.gson.annotations.SerializedName
-import me.igorfedorov.myapp.feature.weather_screen.domain.model.WeatherMain
 
 data class WeatherMainModel(
     @SerializedName("coord")
@@ -15,12 +14,3 @@ data class WeatherMainModel(
     @SerializedName("name")
     val name: String
 )
-
-fun WeatherMainModel.toWeatherMain() =
-    WeatherMain(
-        coordinates = coordinatesModel.toCoordinates(),
-        weather = weatherModel.map { it.toWeather() },
-        base = base,
-        main = mainModel.toMain(),
-        name = name
-    )
